Use Vagrant and nbr2mp4 to convert `.arf` files
This code is intended to utilize Vagrant[1] and Virtualbox[2] to create
an environment in which the nbr2mp4[3] utility can be executed, converting
WebEx recording `.arf` files into `.mp3` and `.mp4` files.

With the necessary packages installed on a users workstation to utilize
Vagrant and Virtualbox, it should be a simple matter of executing
`vagrant up` to stand up an Ubuntu 14.04 VM, configure that system, and
automatically convert any `.arf` files contained in the `arf` directory
into `.mp3` and `.mp4` files which will be placed in the `mp3` and `mp4`
directories. The conversion utility will be executed as part of a
Vagrant provision script, so it's possible to add more `.arf` files
retroactively, and convert those by re-running the provisioning process
via `vagrant provision`.

NOTE: While it was intended to convert the `.arf` files into the `.mp4`
format, that functionality doesn't quite work yet so that has been
disabled. Currently, only `.mp3` files will be created; but there's the
possibility that `.mp4` conversion support will be fixed in a later
commit.

References:

    1. https://www.vagrantup.com/
    2. https://www.virtualbox.org/
    3. https://www.webex.co.uk/support/downloads.html
       ("NBR2MP4 Converter" section)